🚚 File: 구조 변경 및 파일 명 변경
- 페이지별 나누는 작업
- index.tsx를 제거하고 layout으로 변경
- 각 페이지를 생성
- 각 페이지 별 로더 생성
- useLoaderData를 각 로더로 경로 변경
