+++ src/pages/index.ts
@@ -1,0 +1,7 @@
+import { pastLoader } from './Past';
+import { presentLoader } from './Present';
+import { futureLoader } from './Future';
+
+export { default as Layout } from './Layout';
+
+export { pastLoader, presentLoader, futureLoader };
