+++ src/templates/index.ts
@@ -2,5 +2,5 @@ export { default as FutureTemplate } from './FutureTemplate';
 export { default as PastCountTemplate } from './PastCountTemplate';
 export { default as PastActivityTemplate } from './PastActivityTemplate';
 export { default as PastTemplate } from './PastTemplate';
-export { default as Layout } from './Layout';
+export { default as LayoutTemplate } from './LayoutTemplate';
 export { default as PresentTemplate } from './PresentTemplate';
